+++ HISTORY.rst
@@ -10,7 +10,11 @@ History
   * Add support for much slower speed in game widget.
   * Add benchmarks.
   * Add support for setting tasks in environment.
-  * Add wrapper that addd task parameters to observation space.
+  * Add wrappers:
+
+    *  that add task parameters to observation space.
+    *  that clip the reward
+    *  that clip the action space
 
 * Improvements